backend/requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
httpx==0.27.0
python-dotenv==1.0.1
pydantic==2.6.4
pydantic-settings==2.2.1
sqlalchemy==2.0.28
aiosqlite==0.20.0
qdrant-client==1.8.2
pymupdf==1.24.1
python-docx==1.1.0
python-multipart==0.0.9
PyYAML==6.0.1
